feat: refactor search input handling and keyboard navigation in SearchInput and HeaderLeftToolBarInput
- Introduced useSearchSelectedIndex hook to manage keyboard navigation for search results, improving user experience.
- Updated HeaderLeftToolBarInput and SearchInput components to utilize the new hook, streamlining the handling of selected index and key events.
- Enhanced useSearchModalData to calculate total items for better integration with the new navigation logic.
- Cleaned up unused state management and effects, improving code clarity and maintainability.
