+++ packages/kit/src/views/Discovery/hooks/useSearchModalData.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 
 import { useIntl } from 'react-intl';
 
@@ -88,6 +88,20 @@ export function useSearchModalData(searchValue: string) {
     (localData?.bookmarkData ?? []).length > 0 && !displaySearchList;
   const displayHistoryList = (localData?.historyData ?? []).length > 0;
 
+  // Calculate total items
+  const totalItems = useMemo(() => {
+    const searchCount = displaySearchList ? searchList.length : 0;
+    const historyCount = displayHistoryList
+      ? localData?.historyData?.length || 0
+      : 0;
+    return searchCount + historyCount;
+  }, [
+    displaySearchList,
+    searchList.length,
+    displayHistoryList,
+    localData?.historyData?.length,
+  ]);
+
   return {
     localData: localData ?? null,
     refreshLocalData,
@@ -97,5 +111,6 @@ export function useSearchModalData(searchValue: string) {
     displayHistoryList,
     SEARCH_ITEM_ID,
     isEmpty: !displaySearchList && !displayBookmarkList && !displayHistoryList,
+    totalItems,
   };
 }
