+++ packages/kit/src/views/Discovery/pages/Dashboard/Welcome/SearchInput.tsx
@@ -25,6 +25,7 @@ import { shortcutsKeys } from '@onekeyhq/shared/src/shortcuts/shortcutsKeys.enum
 
 import { SearchResultContent } from '../../../components/SearchResultContent';
 import { useSearchModalData } from '../../../hooks/useSearchModalData';
+import { useSearchSelectedIndex } from '../../../hooks/useSearchSelectedIndex';
 
 import { KeyboardShortcutKey } from './KeyboardShortcutKey';
 import { SearchPopover } from './SearchPopover';
@@ -32,13 +33,10 @@ import { SearchPopover } from './SearchPopover';
 import type { ISearchResultContentRef } from '../../../components/SearchResultContent';
 import type { TextInput } from 'react-native';
 
-const ITEM_HEIGHT = 48; // Height of each item in the search results
-
 export function SearchInput() {
   const intl = useIntl();
   const [searchValue, setSearchValue] = useState('');
   const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-  const [selectedIndex, setSelectedIndex] = useState(-1);
   const searchResultRef = useRef<ISearchResultContentRef>(null);
   const scrollViewRef = useRef<IScrollViewRef>(null);
   const inputRef = useRef<TextInput>(null);
@@ -51,10 +49,6 @@ export function SearchInput() {
     }
   }, []);
 
-  useEffect(() => {
-    setSelectedIndex(-1);
-  }, [isPopoverOpen]);
-
   const {
     localData,
     searchList,
@@ -62,13 +56,8 @@ export function SearchInput() {
     displayHistoryList,
     SEARCH_ITEM_ID,
     refreshLocalData,
+    totalItems,
   } = useSearchModalData(searchValue);
-
-  useEffect(() => {
-    scrollViewRef?.current?.scrollTo({
-      y: 0,
-    });
-  }, [searchValue]);
 
   useEffect(() => {
     if (isPopoverOpen) {
@@ -88,7 +77,6 @@ export function SearchInput() {
 
   const handleInputChange = useCallback((text: string) => {
     setSearchValue(text);
-    setSelectedIndex(-1);
   }, []);
 
   const handleInputBlur = useCallback(() => {
@@ -108,67 +96,26 @@ export function SearchInput() {
     }
   });
 
-  useEffect(() => {
-    if (scrollViewRef.current) {
-      const getSelectedItemDistance = () => {
-        if (selectedIndex < 4) return 0;
-
-        // Calculate item height based on your UI design
-        return selectedIndex * ITEM_HEIGHT;
-      };
-
-      const distance = getSelectedItemDistance();
-      scrollViewRef.current.scrollTo({
-        y: distance,
-        animated: true,
-      });
-    }
-  }, [selectedIndex]);
-
-  const handleKeyDown = useCallback(
-    (e: KeyboardEvent) => {
-      // Prevent default behavior for up and down arrow keys
-      if (e.key === 'ArrowUp' || e.key === 'ArrowDown') {
-        e.preventDefault();
-
-        // Calculate total items count
-        const searchCount = displaySearchList ? searchList.length : 0;
-        const historyCount = displayHistoryList
-          ? localData?.historyData?.length || 0
-          : 0;
-        const totalItems = searchCount + historyCount;
-
-        if (totalItems === 0) return;
-
-        // Update selected index based on arrow key
-        if (e.key === 'ArrowDown') {
-          setSelectedIndex((prev) => (prev + 2 > totalItems ? prev : prev + 1));
-        } else if (e.key === 'ArrowUp') {
-          setSelectedIndex((prev) => (prev > -1 ? prev - 1 : -1));
-        }
-      }
-
-      // Handle Enter key press - call openSelectedItem
-      if (e.key === 'Enter') {
-        e.preventDefault();
+  const { selectedIndex, handleKeyDown, resetSelectedIndex } =
+    useSearchSelectedIndex({
+      scrollViewRef,
+      totalItems,
+      searchValue,
+      onEnterPress: () => {
         if (searchResultRef.current) {
           searchResultRef.current.openSelectedItem();
           setIsPopoverOpen(false);
         }
-      }
-
-      if (e.key === 'Escape') {
+      },
+      onEscape: () => {
         setIsPopoverOpen(false);
         inputRef.current?.blur();
-      }
-    },
-    [
-      displaySearchList,
-      searchList.length,
-      displayHistoryList,
-      localData?.historyData?.length,
-    ],
-  );
+      },
+    });
+
+  useEffect(() => {
+    resetSelectedIndex();
+  }, [isPopoverOpen, resetSelectedIndex]);
 
   const isPopoverVisible = useMemo(
     () =>
